refactor: Dashboard周りの修正。AwesomeCountMapperを新設し、各アクションのレコード数取得を集約。同じくAwesomeCountRepository, AwesomeCountServiceも新設し、それぞれDBアクセス、えらいポイントの集計を担当。
