--- src/main/java/webapp/AwesomeCollect/dto/DashboardDto.java
@@ -1,10 +1,0 @@
-package webapp.AwesomeCollect.dto;
-
-import lombok.Data;
-
-@Data
-public class DashboardDto {
-
-  private int totalAwesome;
-  private LearningTimeDto learningTimeDto;
-}
