+++ src/main/java/webapp/AwesomeCollect/dto/dashboard/DashboardDto.java
@@ -1,0 +1,15 @@
+package webapp.AwesomeCollect.dto.dashboard;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+/**
+ * ダッシュボードのデータオブジェクト。
+ */
+@Data
+@AllArgsConstructor
+public class DashboardDto {
+
+  private int totalAwesome;
+  private LearningTimeDto learningTimeDto;
+}
