+++ src/main/java/webapp/AwesomeCollect/dto/dashboard/LearningTimeDto.java
@@ -1,12 +1,15 @@
-package webapp.AwesomeCollect.dto;
+package webapp.AwesomeCollect.dto.dashboard;
 
 import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Data;
-import webapp.AwesomeCollect.entity.AvgLearningTime;
-import webapp.AwesomeCollect.entity.TagLearningTime;
-import webapp.AwesomeCollect.entity.TotalLearningTime;
+import webapp.AwesomeCollect.entity.dashboard.AvgLearningTime;
+import webapp.AwesomeCollect.entity.dashboard.TagLearningTime;
+import webapp.AwesomeCollect.entity.dashboard.TotalLearningTime;
 
+/**
+ * 学習時間のデータオブジェクト。
+ */
 @Data
 @AllArgsConstructor
 public class LearningTimeDto {
