+++ src/main/java/webapp/AwesomeCollect/entity/dashboard/AvgLearningTime.java
@@ -1,8 +1,11 @@
-package webapp.AwesomeCollect.entity;
+package webapp.AwesomeCollect.entity.dashboard;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+/**
+ * 曜日別平均学習時間を扱うオブジェクト。
+ */
 @Data
 @AllArgsConstructor
 public class AvgLearningTime {
