+++ src/main/java/webapp/AwesomeCollect/entity/dashboard/TagLearningTime.java
@@ -1,7 +1,10 @@
-package webapp.AwesomeCollect.entity;
+package webapp.AwesomeCollect.entity.dashboard;
 
 import lombok.Data;
 
+/**
+ * タグ別学習時間を扱うオブジェクト。
+ */
 @Data
 public class TagLearningTime {
   private int tagId;
