+++ src/main/java/webapp/AwesomeCollect/entity/dashboard/TotalLearningTime.java
@@ -1,4 +1,4 @@
-package webapp.AwesomeCollect.entity;
+package webapp.AwesomeCollect.entity.dashboard;
 
 import java.time.LocalDate;
 import lombok.AllArgsConstructor;
