+++ src/main/java/webapp/AwesomeCollect/mapper/dashboard/AwesomeCountMapper.java
@@ -1,4 +1,4 @@
-package webapp.AwesomeCollect.mapper;
+package webapp.AwesomeCollect.mapper.dashboard;
 
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Select;
